:recycle: Refactor Code - Exercício 04
